Queue system created
Created and impremented queu system to hold the entries in the resourcepack folder and the query for the merge function.
Queue system includes moving a entry up the queu when some other is removed.
Moved freeFolder(lang) to the end of the main function again. It should hold the other lang files for switch. Also intend to nest all the functions to prepare the lang file in a function.