scanFolder now takes a target
Updated scanFolder function to take a position of the scan target in RP1 folder. -1 will take the path as the target.
Changed cursor[1] to save the cursor position on the sidebar, cursor[2] now indicates the cursor is on the sidebar or the subwindow
Updated TAB switch case to take the specific window first entry's lenght to highlight
Updated switch cursor[2] case 0 (sidebar navigation) to display the entries of that window in the subwindow