Reestructured input switch and draw switch
Draw switch for sidebar and subwindow separated to accept a update trigger when changin resourepacks.
Resoucepack change tested and working
Input switch reestructures, now cursor[1] is used only to navigate the sidebar and cusor[0] to navigate the subwindow options.
Tab case nows changes the highlight lenght when focusing on the subwindow checking the first entry.
Tab case now checks if the tab has entries, blocking if not.