Windows now are global and can be updated
updateWindows function was created to ease the resize call from any tool function.
Created logger function to concatenatea log string for the program to print in the miniwindow and as a file.
confirmationDialog now has a second type to wait the confirmation of the user, with the purpose of displaing messages.
Fixed "location" string on scanFolder not beeing freed.
translated string was made global