Node comparator function protorype ready
Added a second folder tracking array for the second override, a function will transverse between both folders starting from the trackers and compare the names, entering the matches.
This will result in it entering a branch until there is no coincidences anymore, then the coincident subfolders will have it's contents merged and the exclusive subolders copied, then restarting the cycle leaving the node and going to the end of the next if there is one.